Refactored promotion type enum.
