+++ src/offers/entities/promotion-type.enum.ts
@@ -1,0 +1,4 @@
+export enum PromotionType {
+  LIGHTNING_DEAL = 'LIGHTNING_DEAL',
+  DEAL_OF_THE_DAY = 'DEAL_OF_THE_DAY',
+}
